# noblelift-backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Noblelift App API
  version: 1.0.0
  description: |
    REST API для корпоративной системы Noblelift.
    Согласовано с ТЗ и текущим фронтендом. Таймстемпы — Unix ms (int64), camelCase в JSON.
servers:
  - url: https://api.example.com/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local
...
security:
  - bearerAuth: []
paths:
...
  /tasks:
    get:
      tags: [Tasks]
      summary: Список задач
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - in: query
          name: status
          schema:
            type: array
            items: { $ref: '#/components/schemas/TaskStatusCode' }
          style: form
          explode: false
        - in: query
          name: priority
          schema:
            type: array
            items: { $ref: '#/components/schemas/PriorityCode' }
          style: form
          explode: false
        - in: query
          name: assigneeId
          schema: { type: integer, format: int64 }
        - in: query
          name: creatorId
          schema: { type: integer, format: int64 }
        - in: query
          name: topicId
          schema: { type: integer }
        - in: query
          name: isPrivate
          schema: { type: boolean }
        - in: query
          name: type
          schema: { $ref: '#/components/schemas/TaskType' }
        - in: query
          name: archived
          schema: { type: boolean }
        - in: query
          name: dueFrom
          schema: { type: integer, format: int64, description: 'Unix ms' }
        - in: query
          name: dueTo
          schema: { type: integer, format: int64, description: 'Unix ms' }
        - in: query
          name: q
          schema: { type: string, description: 'Поиск по title/content' }
        - in: query
          name: sort
          schema: { type: string, example: 'dueDate,-priority' }
      responses:
        '200':
          description: List
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedTasks' }
    post:
      tags: [Tasks]
      summary: Создать задачу
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
  /tasks/available:
    get:
      tags: [Tasks]
      summary: Общие задачи (не взятые)
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedTasks' }
...
  /tasks/{id}:
    get:
      tags: [Tasks]
      summary: Карточка задачи
      parameters: [ { $ref: '#/components/parameters/TaskId' } ]
      responses:
        '200':
          description: Task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
    patch:
      tags: [Tasks]
      summary: Обновить свойства задачи
      parameters: [ { $ref: '#/components/parameters/TaskId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
    delete:
      tags: [Tasks]
      summary: Удалить/архивировать задачу (по политике)
      parameters: [ { $ref: '#/components/parameters/TaskId' } ]
      responses:
        '204': { description: No Content }
  /tasks/{id}/assign:
    post:
      tags: [Tasks]
      summary: Назначить исполнителя
      parameters: [ { $ref: '#/components/parameters/TaskId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskAssign' }
      responses:
        '200':
          description: Task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
  /tasks/{id}/unassign:
    post:
      tags: [Tasks]
      summary: Снять исполнителя
      parameters: [ { $ref: '#/components/parameters/TaskId' } ]
      responses:
        '200':
          description: Task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
  /tasks/{id}/take:
    post:
      tags: [Tasks]
      summary: Взять общую задачу себе
      parameters: [ { $ref: '#/components/parameters/TaskId' } ]
      responses:
        '200':
          description: Task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
  /tasks/{id}/release:
    post:
      tags: [Tasks]
      summary: Вернуть задачу в общие
      parameters: [ { $ref: '#/components/parameters/TaskId' } ]
      responses:
        '200':
          description: Task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
...
  /tasks/{id}/archive:
    post:
      tags: [Tasks]
      summary: Архивировать задачу (только если done)
      parameters: [ { $ref: '#/components/parameters/TaskId' } ]
      responses:
        '200':
          description: Task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
  /tasks/{id}/unarchive:
    post:
      tags: [Tasks]
      summary: Вернуть из архива
      parameters: [ { $ref: '#/components/parameters/TaskId' } ]
      responses:
        '200':
          description: Task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      in: query
      name: limit
      schema: { type: integer, default: 20, minimum: 1, maximum: 200 }
    Offset:
      in: query
      name: offset
      schema: { type: integer, default: 0, minimum: 0 }
...
    TaskId:
      in: path
      name: id
      required: true
      schema: { type: integer, format: int64 }
...
  schemas:
...
    TaskType:
      type: string
      enum: [regular, common]
    PriorityCode:
      type: string
      enum: [low, medium, high, urgent]
    Priority:
      type: object
      properties:
        code: { $ref: '#/components/schemas/PriorityCode' }
        label: { type: string, example: 'Средний' }
    TaskStatusCode:
      type: string
      enum: [new, in_progress, pause, done, cancelled]
    TaskStatus:
      type: object
      properties:
        code: { $ref: '#/components/schemas/TaskStatusCode' }
        label: { type: string, example: 'В работе' }
    Task:
      type: object
      properties:
        id: { type: integer, format: int64 }
        creatorId: { type: integer, format: int64 }
        assigneeId: { type: integer, format: int64, nullable: true }
        title: { type: string }
        content: { type: string, nullable: true }
        createdAt: { type: integer, format: int64 }
        updatedAt: { type: integer, format: int64 }
        dueDate: { type: integer, format: int64, nullable: true }
        priority: { $ref: '#/components/schemas/Priority' }
        status: { $ref: '#/components/schemas/TaskStatus' }
        topicId: { type: integer, nullable: true }
        topicText: { type: string, nullable: true }
        isPrivate: { type: boolean }
        type: { $ref: '#/components/schemas/TaskType' }
        archived: { type: boolean }
        attachments:
          type: array
          items: { $ref: '#/components/schemas/TaskFile' }
    TaskCreate:
      type: object
      required: [ title ]
      properties:
        title: { type: string }
        content: { type: string, nullable: true }
        dueDate: { type: integer, format: int64, nullable: true }
        priorityCode: { $ref: '#/components/schemas/PriorityCode' }
        statusCode: { $ref: '#/components/schemas/TaskStatusCode' }
        isPrivate: { type: boolean }
        type: { $ref: '#/components/schemas/TaskType' }
        topicId: { type: integer, nullable: true }
        topicText: { type: string, nullable: true }
        assigneeId: { type: integer, format: int64, nullable: true }
    TaskUpdate:
      type: object
      properties:
        title: { type: string }
        content: { type: string, nullable: true }
        dueDate: { type: integer, format: int64, nullable: true }
        priorityCode: { $ref: '#/components/schemas/PriorityCode' }
        topicId: { type: integer, nullable: true }
        topicText: { type: string, nullable: true }
        isPrivate: { type: boolean }
        type: { $ref: '#/components/schemas/TaskType' }
    TaskAssign:
      type: object
      required: [assigneeId]
      properties:
        assigneeId: { type: integer, format: int64 }
...
    PaginatedTasks:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Task' }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }

    # ----- Task Files / Events -----
    TaskFile:
      type: object
      properties:
        id: { type: integer, format: int64 }
        taskId: { type: integer, format: int64 }
        fileKey: { type: string }
        fileName: { type: string }
        size: { type: integer, nullable: true }
        mime: { type: string, nullable: true }
        uploadedBy: { type: integer, format: int64, nullable: true }
        createdAt: { type: integer, format: int64 }
...